requirements.txt
numpy==2.4.1
pandas==2.3.3
python-dotenv==1.2.1
Requests==2.32.5
scikit_learn==1.8.0
SQLAlchemy==2.0.45
matplotlib==3.10.7
plotly==6.5.2
seaborn==0.13.2